// src/app/shared/model/Request/Product/ProdOfferingObj.model.ts
export class ProdOfferingObj {
    ProdName:any;
    ProdOfferingId: any;
    ProdOfferingCode: any;
    ProdOfferingName: any;
    ProdOfferingDescr: any;
    StartDt: any;
    EndDt: any;
    RowVersion: any;
    ProdOfferingStat :any;
    ProdOfferingHId: any;
    ProdHId: any;
  constructor() { this.ProdOfferingId = 0; this.RowVersion = "" }
}